Refactor configuration structure by consolidating base section logic into a new section module and updating related imports across the codebase.
