+++ libs/bootstrap/app_config.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Literal, Optional, Union
 
-from libs.bootstrap.base_section import BaseSection
+from libs.bootstrap.section import AliasSection, BaseSection, MahjongSection, SettingSection
 from libs.commands.graph.entry import GraphConfig
 from libs.commands.help.entry import HelpConfig
 from libs.commands.ranking.entry import RankingConfig
@@ -19,229 +19,6 @@ from libs.commands.results.entry import ResultsConfig
 from libs.data.lookup import read_memberslist
 from libs.domain.rule import RuleSet
 from libs.types import GradeTableDict
-
-
-class MahjongSection(BaseSection):
-    """mahjongセクション処理"""
-
-    def __init__(self):
-        self.mode: Literal[3, 4] = 4
-        """ 集計モード切替(四人打ち/三人打ち)"""
-        self.rule_version: str = str("default_rule")
-        """ルール識別子"""
-        self.origin_point: int = int(-1)
-        """配給原点"""
-        self.return_point: int = int(-1)
-        """返し点"""
-        self.rank_point: list = []
-        """順位点"""
-        self.ignore_flying: bool = False
-        """トビカウント
-        - *True*: なし
-        - *False*: あり
-        """
-        self.draw_split: bool = False
-        """同点時の順位点
-        - *True*: 山分けにする
-        - *False*: 席順で決める
-        """
-        self.undefined_word = 0
-        """未定義ワードタイプ"""
-
-    def config_load(self, outer: "AppConfig"):
-        """設定値取り込み
-
-        Args:
-            outer (AppConfig): 設定クラスオブジェクト
-        """
-
-        self.section: str = "mahjong"
-        self._parser = outer._parser
-        super().__init__(self)
-
-        # デフォルト値
-        match self.mode:
-            case 3:
-                if self.origin_point < 0:
-                    self.origin_point = 350
-                if self.return_point < 0:
-                    self.return_point = 400
-                if not self.rank_point:
-                    self.rank_point = [30, 0, -30]
-            case 4:
-                if self.origin_point < 0:
-                    self.origin_point = 250
-                if self.return_point < 0:
-                    self.return_point = 300
-                if not self.rank_point:
-                    self.rank_point = [30, 10, -10, -30]
-
-        self.rank_point = self.rank_point[: self.mode]
-
-        logging.debug("%s: %s", self.section, self)
-
-
-class SettingSection(BaseSection):
-    """settingセクション処理"""
-
-    help: str
-    """ヘルプ表示キーワード"""
-    keyword: str
-    """成績記録キーワード(プライマリ)"""
-    remarks_word: str
-    """メモ記録用キーワード"""
-    rule_config: Path
-    """ルール設定ファイル"""
-    default_rule: str
-    """ルール識別子未指定時に使用される識別子"""
-    separate: bool
-    """スコア入力元識別子別集計フラグ
-    - *True*: 識別子別に集計
-    - *False*: すべて集計
-    """
-    channel_id: Optional[str]
-    """チャンネルIDを上書きする"""
-    time_adjust: int
-    """日付変更後、集計範囲に含める追加時間"""
-    search_word: str
-    """コメント固定(検索時の検索文字列)"""
-    group_length: int
-    """コメント固定(検索時の集約文字数)"""
-    guest_mark: str
-    """ゲスト無効時に未登録メンバーに付与する印"""
-    database_file: Union[str, Path]
-    """成績管理データベースファイル名"""
-    backup_dir: Optional[Path]
-    """バックアップ先ディレクトリ"""
-    font_file: Path
-    """グラフ描写に使用するフォントファイル"""
-    graph_style: str
-    """グラフスタイル"""
-    work_dir: Path
-
-    def __init__(self):
-        self._reset()
-
-    def _reset(self):
-        self.help = str("麻雀成績ヘルプ")
-        self.keyword = str("終局")
-        self.remarks_word = str("麻雀成績メモ")
-        self.rule_config = Path("files/default_rule.ini")
-        self.time_adjust = int(12)
-        self.default_rule = str("")
-        self.separate = bool(False)
-        self.channel_id = None
-        self.search_word = str("")
-        self.group_length = int(0)
-        self.guest_mark = str("※")
-        self.database_file = Path("mahjong.db")
-        self.backup_dir = None
-        self.font_file = Path("ipaexg.ttf")
-        self.graph_style = str("ggplot")
-        self.work_dir = Path("work")
-
-    def config_load(self, outer: "AppConfig"):
-        """設定値取り込み
-
-        Args:
-            outer (AppConfig): 設定クラスオブジェクト
-        """
-
-        self.section: str = "setting"
-        self._parser = outer._parser
-        self._reset()
-        super().__init__(self)
-
-        # 成績登録キーワード
-        if not (isinstance(self.keyword, Path) and self.keyword.exists()):
-            self.keyword = str(self.keyword)
-
-        # デフォルトルール識別子
-        if not self.default_rule:
-            self.default_rule = outer.mahjong.rule_version
-
-        # フォントファイルチェック
-        for chk_dir in (outer.config_dir, outer.script_dir):
-            chk_file = chk_dir / str(self.font_file)
-            if chk_file.exists():
-                self.font_file = chk_file
-                break
-        else:
-            if not self.font_file.exists():
-                logging.critical("The specified font file cannot be found.")
-                sys.exit(255)
-
-        # 作業ディレクトリパス
-        if not self.work_dir.is_absolute():
-            self.work_dir = outer.script_dir / self.work_dir
-
-        # データベース関連
-        if isinstance(self.database_file, Path) and not self.database_file.exists():
-            self.database_file = outer.config_dir / str(self.database_file)
-
-        logging.debug("%s: %s", self.section, self)
-
-
-class AliasSection(BaseSection):
-    """aliasセクション処理"""
-
-    results: list
-    """成績サマリ出力コマンド"""
-    graph: list
-    """成績グラフ出力コマンド"""
-    ranking: list
-    """ランキング出力コマンド"""
-    report: list
-    """レポート出力コマンド"""
-    download: list
-    member: list
-    """メンバーリスト表示コマンド"""
-    add: list
-    delete: list
-    team_create: list
-    team_del: list
-    team_add: list
-    team_remove: list
-    team_list: list
-    """チームリスト出力コマンド"""
-    team_clear: list
-
-    def __init__(self):
-        self._reset()
-
-    def _reset(self):
-        self.results = ["results", "成績"]
-        self.graph = ["graph", "グラフ"]
-        self.ranking = ["ranking", "ランキング"]
-        self.report = ["report", "レポート"]
-        self.download = ["download", "ダウンロード"]
-        self.member = ["member", "userlist", "member_list"]
-        self.add = ["add"]
-        self.delete = ["del"]
-        self.team_create = ["team_create"]
-        self.team_del = ["team_del"]
-        self.team_add = ["team_add"]
-        self.team_remove = ["team_remove"]
-        self.team_list = ["team_list"]
-        self.team_clear = ["team_clear"]
-
-    def config_load(self, outer: "AppConfig"):
-        """設定値取り込み
-
-        Args:
-            outer (AppConfig): 設定クラスオブジェクト
-        """
-
-        self.section: str = "alias"
-        self._parser = outer._parser
-        self._reset()
-        super().__init__(self)
-
-        # delのエイリアス取り込み(設定ファイルに`delete`と書かれていない)
-        list_data = [x.strip() for x in str(self._parser.get("alias", "del", fallback="del")).split(",")]
-        self.delete.extend(list_data)
-
-        logging.debug("%s: %s", self.section, self)
 
 
 class DropItems(BaseSection):
