+++ libs/bootstrap/configuration.py
@@ -19,6 +19,7 @@ import libs.commands.results.entry
 import libs.global_value as g
 from integrations import factory
 from libs.bootstrap.app_config import AppConfig
+from libs.bootstrap.section import SubCommands
 from libs.commands.registry import member, team
 from libs.data import initialization, lookup
 from libs.functions.compose import text_item
@@ -26,7 +27,6 @@ from libs.types import Args, StyleOptions
 
 if TYPE_CHECKING:
     from integrations.protocols import MessageParserProtocol
-    from libs.types import SubCommandsConfigType
 
 
 def set_loglevel():
@@ -347,16 +347,13 @@ def register():
     for command, ep in dispatch_table.items():
         # 呼び出しキーワード登録
         if hasattr(g.cfg, command):
-            sub_command = cast("SubCommandsConfigType", getattr(g.cfg, command))
-            if hasattr(sub_command, "default_commandword"):
-                print(">", command, sub_command.default_commandword, sub_command.commandword)
-            for commandword in sub_command.commandword:
-                g.keyword_dispatcher.update({commandword: ep})
-        # コマンドサフィックス登録
-        if hasattr(sub_command, "command_suffix"):
-            for keyword in g.cfg.rule.keyword_mapping:
-                for suffix in sub_command.command_suffix:
-                    g.keyword_dispatcher.update({f"{keyword}{suffix}": ep})
+            sub_command = getattr(g.cfg, command)
+            if isinstance(sub_command, SubCommands):
+                for commandword in sub_command.commandword:
+                    g.keyword_dispatcher.update({commandword: ep})
+                for keyword in g.cfg.rule.keyword_mapping:  # コマンドサフィックス登録
+                    for suffix in sub_command.command_suffix:
+                        g.keyword_dispatcher.update({f"{keyword}{suffix}": ep})
         # スラッシュコマンド登録
         if hasattr(g.cfg.alias, command):
             for alias in cast(list, getattr(g.cfg.alias, command)):
